+++ Avatar 2.0 VRCSDK/Dependencies/VRChat/Editor/ControlPanel/TBASdkChangelog.cs
@@ -69,6 +69,7 @@ UnityEditor.EditorPrefs.GetFloat("SDKColor_A")
         EditorGUILayout.LabelField("ODDS (SDX is based in ODDS)", EditorStyles.boldLabel);
         EditorGUILayout.LabelField("Plague (LoadBundle Script Creator)", EditorStyles.boldLabel);
         EditorGUILayout.LabelField("Controser (AntiYoink[In 3.0] and SENTINEL IMPORTER)", EditorStyles.boldLabel);
+        EditorGUILayout.LabelField("Bunny (For their Utilities!)", EditorStyles.boldLabel);
         EditorGUILayout.EndVertical();
 
         changesScroll = EditorGUILayout.BeginScrollView(changesScroll, GUILayout.Width(SdkWindowWidth));
@@ -84,6 +85,20 @@ Much easier to do than to make an entire remake
 This is completely free of charge as I refuse to make
 Any money or profit from my SDX and I will not rebrand
 My SDX, this decision is final and wont be changed.
+
+The Black Arms SDX 2.1
+Added several utilities (thanks to bunny!)
+-AddVRCAvatarPedestals
+-ChangePropertiesOfProject
+-CopyBones
+-CopyComponents
+-Duplicator
+-FindMaterial
+-FindWrongScripts
+-GetPropertiesOfObject
+-LoadBundle (Unreal Loader)
+-MeshGenerator
+-MeshToAsset
 
 The Black Arms SDX 2.0.3
 TBA got banned again so I updated links